// app/lib/types.ts
export type Artist = {
  //name?: string | undefined; // Fixed syntax error here
  name:string;
  picture?: string;
  info?: string;
  id: number|string | null;
  image?: string;
  gender?: string;
  followers?: number;
  monthlyListeners?: number;
  topSong?: string;
  genre?: string| string[];
}




// Main Track interface based on usage in the code
export type Track = {
  id: number;
  track_name: string;
  music_id?:string;
  genre: string;
  duration: string;
  label: string;
  image: string;
  actor?:string;
  artist?:Artist|undefined;
  plays: number;
 url?: string;
  //releaseDate?: date;
  releaseDate?: string;
  isLiked?: boolean;
  audio?: string;
  no_plays?:number;
  description?: string;
  info?: string;
}
// ...
export type  HeroSlide = {

  id: string;
  title: string;
  artist: string;
  shuffle: boolean;
  repeat: 'off' | 'one' | 'all';
   queue: Track[];
  currentIndex: number;
  image: string;
  type: 'song' | 'album' | 'artist';
  plays?: number;
  duration: string;
  releaseDate?: string;
  track:Track;
  
}

// News Item Type Definition
export type NewsItem = {
  id: string;
  title: string;
  image: string;
  type?: 'event' | 'release' | 'news' | 'interview';
  date: string;
  link: string;
}